+++ extensions/thock/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Thock Changelog
 
+## [Preview Soundpack] - {PR_MERGE_DATE}
+
+- Added a `Preview Soundpack` command that lists the soundpack registry and plays a few keystrokes from any pack without installing it
+- Installed packs are tagged and can be set as the active soundpack from the same list
+
 ## [Chore: Updated the README] - 2025-05-12
 
 ## [Initial Version] - 2025-05-02